apply new UI
