+++ app/build.gradle
@@ -11,6 +11,7 @@ android {
         versionName "1.0"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
         multiDexEnabled true
+        vectorDrawables.useSupportLibrary = true
     }
     buildTypes {
         release {
@@ -42,6 +43,8 @@ dependencies {
 
     implementation 'com.mlsdev.rximagepicker:library:2.0.0'
     implementation 'com.github.bumptech.glide:glide:4.0.0'
+    implementation 'com.android.support:support-vector-drawable:27.0.2'
+    implementation 'com.android.support:support-v4:27.0.2'
     annotationProcessor 'com.github.bumptech.glide:compiler:4.0.0'
 
     implementation group: 'com.google.code.gson', name: 'gson', version: '2.3.1'
@@ -52,6 +55,12 @@ dependencies {
     implementation 'com.squareup.retrofit2:retrofit:2.2.0'
     implementation 'com.squareup.retrofit2:converter-gson:+'
 }
+
+
+
+
+
+
