+++ app/src/main/java/com/gomsang/lab/publicchain/ui/fragments/navigations/MoreFragment.java
@@ -1,0 +1,35 @@
+package com.gomsang.lab.publicchain.ui.fragments.navigations;
+
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.gomsang.lab.publicchain.R;
+
+public class MoreFragment extends Fragment {
+
+    public MoreFragment() {
+        // Required empty public constructor
+    }
+
+    public static MoreFragment newInstance() {
+        MoreFragment fragment = new MoreFragment();
+        Bundle args = new Bundle();
+        fragment.setArguments(args);
+        return fragment;
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_more, container, false);
+    }
+}
