+++ app/src/main/res/layout/app_bar_main2.xml
@@ -5,7 +5,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="com.gomsang.lab.publicchain.ui.activities.MainActivity">
+        tools:context="com.gomsang.lab.publicchain.ui.activities.Main1Activity">
 
         <android.support.design.widget.AppBarLayout
             android:layout_width="match_parent"
@@ -23,7 +23,7 @@
 
         <include
             android:id="@+id/contentPanel"
-            layout="@layout/content_main" />
+            layout="@layout/content_main2" />
 
         <android.support.design.widget.FloatingActionButton
             android:id="@+id/fab"
