+++ app/src/main/res/layout/content_main2.xml
@@ -7,7 +7,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         app:layout_behavior="@string/appbar_scrolling_view_behavior"
-        tools:context="com.gomsang.lab.publicchain.ui.activities.MainActivity">
+        tools:context="com.gomsang.lab.publicchain.ui.activities.Main1Activity">
 
         <fragment
             android:id="@+id/contentFragment"
